pyproject.toml
[project]
name = "wave2-vader"
version = "0.1.0"
description = "Background Detection"
authors = [
    {name = "Joscha Bisping",email = "joscha.l.bisping@tu-berlin.de"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "jupyter (>=1.1.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "colour-science (>=0.4.6,<0.5.0)",
    "lightning (>=2.5.1.post0,<3.0.0)",
    "transformers (>=4.52.4,<5.0.0)",
    "accelerate (>=1.7.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

